part_3/dogs_api/requirements.txt
Django == 5.2.1
djangorestframework == 3.16.0
django-filter == 2.4.0
gunicorn == 21.2.0
psycopg2 == 2.9.10
pylint == 3.3.7
